// src/types/PropertyType.ts
export type PropertyType = {
  _id?: string;
  AGENT_REF: string;
  ADDRESS_1: string;
  ADDRESS_2: string;
  ADDRESS_3: string;
  TOWN: string;
  POSTCODE1: string;
  POSTCODE2: string;
  FEATURE1: string;
  FEATURE2: string;
  FEATURE3: string;
  FEATURE4: string;
  FEATURE5: string;
  FEATURE6: string;
  FEATURE7: string;
  FEATURE8: string;
  FEATURE9: string;
  FEATURE10: string;
  SUMMARY: string;
  DESCRIPTION: string;
  BRANCH_ID: string;
  STATUS_ID: string;
  BEDROOMS: string;
  BATHROOMS: string;
  LIVING_ROOMS: string;
  PRICE: string;
  PRICE_QUALIFIER: string;
  PROP_SUB_ID: string;
  CREATE_DATE: string;
  UPDATE_DATE: string;
  DISPLAY_ADDRESS: string;
  PUBLISHED_FLAG: string;
  LET_DATE_AVAILABLE: string;
  LET_BOND: string;
  LET_TYPE_ID: string;
  LET_FURN_ID: string;
  LET_RENT_FREQUENCY: string;
  LET_WASHING_MACHINE_FLAG: string;
  LET_DISHWASHER_FLAG: string;
  LET_BURGLAR_ALARM_FLAG: string;
  LET_BILL_INC_WATER: string;
  LET_BILL_INC_GAS: string;
  LET_BILL_INC_ELECTRICITY: string;
  LET_BILL_INC_TV_LICENCE: string;
  LET_BILL_INC_TV_SUBSCRIPTION: string;
  LET_BILL_INC_INTERNET: string;
  TENURE_TYPE_ID: string;
  TRANS_TYPE_ID: string;
  NEW_HOME_FLAG: string;
  MEDIA_IMAGE_00: string;
  MEDIA_IMAGE_01: string;
  MEDIA_IMAGE_TEXT_01: string;
  MEDIA_IMAGE_02: string;
  MEDIA_IMAGE_TEXT_02: string;
  MEDIA_IMAGE_03: string;
  MEDIA_IMAGE_TEXT_03: string;
  MEDIA_IMAGE_04: string;
  MEDIA_IMAGE_TEXT_04: string;
  MEDIA_IMAGE_05: string;
  MEDIA_IMAGE_TEXT_05: string;
  MEDIA_IMAGE_06: string;
  MEDIA_IMAGE_TEXT_06: string;
  MEDIA_IMAGE_07: string;
  MEDIA_IMAGE_TEXT_07: string;
  MEDIA_IMAGE_08: string;
  MEDIA_IMAGE_TEXT_08: string;
  MEDIA_IMAGE_09: string;
  MEDIA_IMAGE_TEXT_09: string;
  MEDIA_IMAGE_10: string;
  MEDIA_IMAGE_TEXT_10: string;
  MEDIA_IMAGE_11: string;
  MEDIA_IMAGE_TEXT_11: string;
  MEDIA_IMAGE_12: string;
  MEDIA_IMAGE_TEXT_12: string;
  MEDIA_IMAGE_13: string;
  MEDIA_IMAGE_TEXT_13: string;
  MEDIA_IMAGE_14: string;
  MEDIA_IMAGE_TEXT_14: string;
  MEDIA_IMAGE_15: string;
  MEDIA_IMAGE_TEXT_15: string;
  MEDIA_IMAGE_16: string;
  MEDIA_IMAGE_TEXT_16: string;
  MEDIA_IMAGE_60: string;
  MEDIA_IMAGE_TEXT_60: string;
  MEDIA_IMAGE_61: string;
  MEDIA_IMAGE_TEXT_61: string;
  MEDIA_FLOOR_PLAN_00: string;
  MEDIA_FLOOR_PLAN_TEXT_00: string;
  MEDIA_FLOOR_PLAN_01: string;
  MEDIA_FLOOR_PLAN_TEXT_01: string;
  UNIPOL_STANDARD: string;
  PROPERTY_WEEK: string;
  TENANT_TYPE: string;
  FEED_SLIDE_01: string;
  FEED_SLIDE_02: string;
  FEED_SLIDE_03: string;
  FEED_SLIDE_04: string;
  FEED_SLIDE_05: string;
  FEED_SLIDE_06: string;
  FEED_SLIDE_07: string;
  FEED_SLIDE_08: string;
  FEED_SLIDE_09: string;
  FEED_SLIDE_10: string;
  FEED_SLIDE_11: string;
  FEED_SLIDE_12: string;
  FEED_SLIDE_13: string;
  FEED_SLIDE_14: string;
  FEED_SLIDE_15: string;
  FEED_SLIDE_16: string;
  MAIN_VIEWER: string;
  EPC_EER: string;
  EPC_EERP: string;
  EPC_EIR: string;
  EPC_EIRP: string;
  TLS_INCENTIVE: string;
  TLS_SOCIALMEDIA: string;
  "~": string;
};